Update memory.cpp
According to "https://pasti.fxatari.com/68kdocs/68kPrefetch.html"

"During the first stage extra fetch cycles are performed, one for each extension word. Most of the extension words are processed directly by the instruction and not transferred to the prefetch queue. Otherwise extension words would reach the decoder. Only the last word fetched at this stage is transferred to IRC. This last word is actually not part of the current instruction, it is instead the opcode of the next one. Remember that one extension word was already fetched during the previous instruction."

In previous code extension already fetched in irc was transfered in  ir (prefectch queue) which is incorrect. I don't think it changes the core behavior, it's just more correct ;-)